+++ src/plonetheme/tokyo/tests/test_setup.py
@@ -1,8 +1,6 @@
 # -*- coding: utf-8 -*-
 
 from plone import api
-from plone.app.testing import setRoles
-from plone.app.testing import TEST_USER_ID
 from plonetheme.tokyo.testing import PLONETHEME_TOKYO_INTEGRATION_TESTING
 
 import unittest
@@ -20,13 +18,20 @@ class TestSetup(unittest.TestCase):
 
     def test_product_installed(self):
         """Test if plonetheme.tokyo is installed."""
-        # self.assertTrue(self.installer.isProductInstalled('plonetheme.tokyo'))
+        self.assertTrue(self.installer.isProductInstalled('plonetheme.tokyo'))
 
     def test_browserlayer(self):
         """Test that IPlonethemeTokyoLayer is registered."""
-        # from plonetheme.tokyo.interfaces import IPlonethemeTokyoLayer
-        # from plone.browserlayer import utils
-        # self.assertIn(IPlonethemeTokyoLayer, utils.registered_layers())
+        from plonetheme.tokyo.interfaces import IPlonethemeTokyoLayer
+        from plone.browserlayer import utils
+        self.assertIn(IPlonethemeTokyoLayer, utils.registered_layers())
+
+    def testCurrentTheme(self):
+        from plone.app.theming.interfaces import IThemeSettings
+        self.assertEqual(
+            api.portal.get_registry_record('currentTheme', IThemeSettings),
+            'plonetheme.tokyo',
+        )
 
 
 class TestUninstall(unittest.TestCase):
@@ -36,17 +41,22 @@ class TestUninstall(unittest.TestCase):
     def setUp(self):
         self.portal = self.layer['portal']
         self.installer = api.portal.get_tool('portal_quickinstaller')
-        roles_before = api.user.get_roles(TEST_USER_ID)
-        setRoles(self.portal, TEST_USER_ID, ['Manager'])
-        # self.installer.uninstallProducts(['plonetheme.tokyo'])
-        setRoles(self.portal, TEST_USER_ID, roles_before)
+        with api.env.adopt_roles(['Manager']):
+            self.installer.uninstallProducts(['plonetheme.tokyo'])
 
     def test_product_uninstalled(self):
         """Test if plonetheme.tokyo is cleanly uninstalled."""
-        # self.assertFalse(self.installer.isProductInstalled('plonetheme.tokyo'))
+        self.assertFalse(self.installer.isProductInstalled('plonetheme.tokyo'))
 
     def test_browserlayer_removed(self):
-        """Test that IPlonethemeTokyoLayer is removed."""
-        # from plonetheme.tokyo.interfaces import IPlonethemeTokyoLayer
-        # from plone.browserlayer import utils
-        # self.assertNotIn(IPlonethemeTokyoLayer, utils.registered_layers())
+        """ Test that IPlonethemeTokyoLayer is removed."""
+        from plonetheme.tokyo.interfaces import IPlonethemeTokyoLayer
+        from plone.browserlayer import utils
+        self.assertNotIn(IPlonethemeTokyoLayer, utils.registered_layers())
+
+    def testCurrentTheme(self):
+        from plone.app.theming.interfaces import IThemeSettings
+        self.assertEqual(
+            api.portal.get_registry_record('currentTheme', IThemeSettings),
+            'barceloneta',
+        )